[detect] Prevent repatching in various scenarios.
First, we check for the existence of OpenCore-Legacy-Patcher.plist.

If the file does not exist, checking for "Broken" in the diskutil
output for the "Sealed" key on the root volume will help continue -
if it's broken, unpatching is required to continue.

If the file does exist, checking for "Commit URL" will help determine
if the current copy of OCLP was used to install any patches. Should
the patches be from a different commit, unpatching is required to
continue.

With the commit matching, we then check for any non-wireless patches
that are currently installed on the system. This allows for users to
continue patching wireless first and everything else later, while also
ensuring that repatching is blocked when all patches are installed.